Show push permission on all pages
