+++ app/layout.jsx
@@ -1,6 +1,7 @@
 import { Poppins } from "next/font/google";
 import "@/app/globals.css";
 import { Toaster } from "sonner";
+import PushInit from "@/components/PushInit";
 
 const poppins = Poppins({
   weight: ["100", "200", "300", "400", "500", "600", "700", "800", "900"], // choose weights you want
@@ -145,6 +146,7 @@ export default function RootLayout({ children }) {
         className={`${poppins.variable} antialiased scroll font-poppins group overflow-x-hidden overflow-y-auto has-[.leadPopup:checked]:overflow-hidden`}
       >
         <Toaster position="bottom-center" richColors />
+        <PushInit />
         {children}
       </body>
     </html>
