--- app/pwa/loading.jsx
@@ -1,5 +1,0 @@
-import PwaSkeleton from '@/components/skeleton/pwa-skeleton'
-
-export default function Loading() {
-  return <PwaSkeleton />
-}
